+++ max_div/benchmark/sample_int.py
@@ -1,13 +1,14 @@
-import math
-
 import numpy as np
 
 from max_div.internal.benchmarking import BenchmarkResult, benchmark
 from max_div.internal.compat import is_numba_installed
+from max_div.internal.formatting import md_bold, md_italic, md_multiline
 from max_div.sampling.discrete import sample_int
 
+from ._formatting import format_as_markdown, format_for_console
 
-def benchmark_sample_int(turbo: bool = False) -> None:
+
+def benchmark_sample_int(turbo: bool = True, markdown: bool = False) -> None:
     """
     Benchmarks the `sample_int` function from `max_div.sampling.discrete`.
 
@@ -21,6 +22,7 @@ def benchmark_sample_int(turbo: bool = False) -> None:
         * (10000, 10), (10000, 100), (10000, 1000), (10000, 5000), (10000, 10000)
 
     :param turbo: If `True`, a much shorter (but less reliable) benchmark is run; intended for testing purposes.
+    :param markdown: If `True`, outputs the results as a Markdown table.
     """
 
     if not is_numba_installed():
@@ -30,56 +32,66 @@ def benchmark_sample_int(turbo: bool = False) -> None:
 
     print("Benchmarking `sample_int`...")
     print()
-    print("".ljust(30) + "use_numba=False".rjust(25) + "use_numba=True".rjust(25))
 
     for replace, use_p, desc in [
-        (True, False, "with replacement, uniform"),
-        (False, False, "without replacement, uniform"),
-        (True, True, "with replacement, non-uniform"),
-        (False, True, "without replacement, non-uniform"),
+        (True, False, "with replacement, uniform probabilities"),
+        (False, False, "without replacement, uniform probabilities"),
+        (True, True, "with replacement, custom probabilities"),
+        (False, True, "without replacement, custom probabilities"),
     ]:
-        print(desc.upper())
+        print(desc.upper() + ": ", end="")
 
-        for n, k in [
-            (10, 1),
-            (100, 1),
-            (1000, 1),
-            (5000, 1),
-            (10000, 1),
-            (10000, 10),
-            (10000, 100),
-            (10000, 1000),
-            (10000, 5000),
-            (10000, 10000),
-        ]:
-            size_str = f"{k:<6_} out of {n:<6_}"
+        # --- create headers ------------------------------
+        if markdown:
+            headers = [
+                "`k`",
+                "`n`",
+                md_multiline([md_bold("accelerated=False"), md_italic("(numpy)")]),
+                md_multiline([md_bold("accelerated=True"), md_italic("(custom numba)")]),
+            ]
+        else:
+            headers = ["k", "n", "accelerated=False", "accelerated=True"]
 
-            results: list[BenchmarkResult] = []
-            for use_numba in [False, True]:
-                if use_p:
-                    p = np.random.rand(n)
-                    p /= p.sum()
-                else:
-                    p = None
+        # --- benchmark ------------------------------------
+        data: list[list[str | BenchmarkResult]] = []
+        for n in [10, 100, 1000, 10000]:
+            for k in [1, 10, 100, 1000, 10000]:
+                if (not replace) and (k > n):
+                    continue  # skip this combination, since it's not feasible
 
-                def func_to_benchmark():
-                    sample_int(n=n, k=k, replace=replace, p=p, use_numba=use_numba)
+                data_row: list[str | BenchmarkResult] = [str(k), str(n)]
 
-                results.append(
-                    benchmark(
-                        f=func_to_benchmark,
-                        t_per_run=0.001 if turbo else 0.1,
-                        n_warmup=3 if turbo else 10,
-                        n_benchmark=3 if turbo else 30,
-                        silent=True,
+                for accelerated in [False, True]:
+                    if use_p:
+                        p = np.random.rand(n)
+                        p /= p.sum()
+                    else:
+                        p = None
+
+                    def func_to_benchmark():
+                        sample_int(n=n, k=k, replace=replace, p=p, accelerated=accelerated)
+
+                    data_row.append(
+                        benchmark(
+                            f=func_to_benchmark,
+                            t_per_run=0.001 if turbo else 0.1,
+                            n_warmup=3 if turbo else 10,
+                            n_benchmark=3 if turbo else 30,
+                            silent=True,
+                        )
                     )
-                )
 
-            print(
-                (" " * 4)
-                + size_str.ljust(26)
-                + results[0].t_sec_with_uncertainty_str.rjust(25)
-                + results[1].t_sec_with_uncertainty_str.rjust(25)
-            )
+                data.append(data_row)
+                print(".", end="")  # minimalistic progress indicator
+        print()
 
+        # --- show results -----------------------------------------
+        if markdown:
+            display_data = format_as_markdown(headers, data)
+        else:
+            display_data = format_for_console(headers, data)
+
+        print()
+        for line in display_data:
+            print(line)
         print()