+++ tests/benchmark/test_sample_int.py
@@ -1,5 +1,8 @@
+import pytest
+
 from max_div.benchmark import benchmark_sample_int
 
 
-def test_benchmark_sample_int():
-    benchmark_sample_int(turbo=True)
+@pytest.mark.parametrize("markdown", [True, False])
+def test_benchmark_sample_int(markdown: bool):
+    benchmark_sample_int(turbo=True, markdown=markdown)