+++ tests/sampling/discrete/test_sample_int.py
@@ -22,31 +22,31 @@ def get_probabilities(n: int) -> np.ndarray[float]:
 def test_sample_int_argument_validation(use_numba: bool):
     # --- n < 1 ------------------------------------------
     with pytest.raises(ValueError):
-        sample_int(n=0, use_numba=use_numba)
+        sample_int(n=0, accelerated=use_numba)
 
     with pytest.raises(ValueError):
-        sample_int(n=-10, use_numba=use_numba)
+        sample_int(n=-10, accelerated=use_numba)
 
     # --- k < 1 ------------------------------------------
     with pytest.raises(ValueError):
-        sample_int(n=10, k=0, use_numba=use_numba)
+        sample_int(n=10, k=0, accelerated=use_numba)
 
     with pytest.raises(ValueError):
-        sample_int(n=10, k=-5, use_numba=use_numba)
+        sample_int(n=10, k=-5, accelerated=use_numba)
 
     # --- k > n when replace=False -----------------------
     with pytest.raises(ValueError):
-        sample_int(n=10, k=20, replace=False, use_numba=use_numba)
+        sample_int(n=10, k=20, replace=False, accelerated=use_numba)
 
     with pytest.raises(ValueError):
-        sample_int(n=100, k=101, replace=False, use_numba=use_numba)
+        sample_int(n=100, k=101, replace=False, accelerated=use_numba)
 
     # --- p invalid shape --------------------------------
     with pytest.raises(ValueError):
-        sample_int(n=10, k=5, replace=True, p=np.array([0.1, 0.9]), use_numba=use_numba)  # wrong size
+        sample_int(n=10, k=5, replace=True, p=np.array([0.1, 0.9]), accelerated=use_numba)  # wrong size
 
     with pytest.raises(ValueError):
-        sample_int(n=4, k=2, replace=True, p=np.array([[0.1, 0.4], [0.4, 0.1]]), use_numba=use_numba)  # wrong shape
+        sample_int(n=4, k=2, replace=True, p=np.array([[0.1, 0.4], [0.4, 0.1]]), accelerated=use_numba)  # wrong shape
 
 
 # =================================================================================================
@@ -56,7 +56,7 @@ def test_sample_int_argument_validation(use_numba: bool):
 @pytest.mark.parametrize("use_numba", [True, False])
 def test_sample_int_uniform_with_replacement_invariants_single(n: int, use_numba: bool):
     # --- act ---------------------------------------------
-    sample = sample_int(n, replace=True, use_numba=use_numba)
+    sample = sample_int(n, replace=True, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(sample, numbers.Integral)
@@ -79,7 +79,7 @@ def test_sample_int_uniform_with_replacement_invariants_single(n: int, use_numba
 @pytest.mark.parametrize("use_numba", [True, False])
 def test_sample_int_uniform_with_replacement_invariants_multiple(n: int, k: int, use_numba: bool):
     # --- act ---------------------------------------------
-    samples = sample_int(n, k, replace=True, use_numba=use_numba)
+    samples = sample_int(n, k, replace=True, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(samples, np.ndarray)
@@ -91,7 +91,7 @@ def test_sample_int_uniform_with_replacement_invariants_multiple(n: int, k: int,
 @pytest.mark.parametrize("use_numba", [True, False])
 def test_sample_int_uniform_with_replacement_duplicates(use_numba: bool):
     # --- act ---------------------------------------------
-    samples = sample_int(n=1000, k=900, replace=True, use_numba=use_numba)  # very unlikely all are unique
+    samples = sample_int(n=1000, k=900, replace=True, accelerated=use_numba)  # very unlikely all are unique
 
     # --- assert ------------------------------------------
     assert len(set(samples)) < 900  # there are duplicates, which is expected with replacement
@@ -113,9 +113,9 @@ def test_sample_int_uniform_with_replacement_duplicates(use_numba: bool):
 @pytest.mark.parametrize("use_numba", [True, False])
 def test_sample_int_uniform_with_replacement_seed(n: int, k: int, use_numba: bool):
     # --- act ---------------------------------------------
-    samples_1 = sample_int(n, k, replace=True, use_numba=use_numba, seed=123)
-    samples_2 = sample_int(n, k, replace=True, use_numba=use_numba, seed=123)
-    samples_3 = sample_int(n, k, replace=True, use_numba=use_numba, seed=None)
+    samples_1 = sample_int(n, k, replace=True, accelerated=use_numba, seed=123)
+    samples_2 = sample_int(n, k, replace=True, accelerated=use_numba, seed=123)
+    samples_3 = sample_int(n, k, replace=True, accelerated=use_numba, seed=None)
 
     # --- assert ------------------------------------------
     np.testing.assert_array_equal(samples_1, samples_2)
@@ -132,7 +132,7 @@ def test_sample_int_uniform_with_replacement_seed(n: int, k: int, use_numba: boo
 @pytest.mark.parametrize("use_numba", [True, False])
 def test_sample_int_uniform_without_replacement_invariants_single(n: int, use_numba: bool):
     # --- act ---------------------------------------------
-    sample = sample_int(n, replace=False, use_numba=use_numba)
+    sample = sample_int(n, replace=False, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(sample, numbers.Integral)
@@ -154,7 +154,7 @@ def test_sample_int_uniform_without_replacement_invariants_single(n: int, use_nu
 @pytest.mark.parametrize("use_numba", [True, False])
 def test_sample_int_uniform_without_replacement_invariants_multiple(n: int, k: int, use_numba: bool):
     # --- act ---------------------------------------------
-    samples = sample_int(n, k, replace=False, use_numba=use_numba)
+    samples = sample_int(n, k, replace=False, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(samples, np.ndarray)
@@ -179,9 +179,9 @@ def test_sample_int_uniform_without_replacement_invariants_multiple(n: int, k: i
 @pytest.mark.parametrize("use_numba", [True, False])
 def test_sample_int_uniform_without_replacement_seed(n: int, k: int, use_numba: bool):
     # --- act ---------------------------------------------
-    samples_1 = sample_int(n, k, replace=False, use_numba=use_numba, seed=123)
-    samples_2 = sample_int(n, k, replace=False, use_numba=use_numba, seed=123)
-    samples_3 = sample_int(n, k, replace=False, use_numba=use_numba, seed=None)
+    samples_1 = sample_int(n, k, replace=False, accelerated=use_numba, seed=123)
+    samples_2 = sample_int(n, k, replace=False, accelerated=use_numba, seed=123)
+    samples_3 = sample_int(n, k, replace=False, accelerated=use_numba, seed=None)
 
     # --- assert ------------------------------------------
     np.testing.assert_array_equal(samples_1, samples_2)
@@ -201,7 +201,7 @@ def test_sample_int_non_uniform_with_replacement_invariants_single(n: int, use_n
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    sample = sample_int(n, replace=True, p=p, use_numba=use_numba)
+    sample = sample_int(n, replace=True, p=p, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(sample, numbers.Integral)
@@ -227,7 +227,7 @@ def test_sample_int_non_uniform_with_replacement_invariants_multiple(n: int, k: 
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    samples = sample_int(n, k, replace=True, p=p, use_numba=use_numba)
+    samples = sample_int(n, k, replace=True, p=p, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(samples, np.ndarray)
@@ -242,7 +242,7 @@ def test_sample_int_non_uniform_with_replacement_duplicates(use_numba: bool):
     p = get_probabilities(1000)
 
     # --- act ---------------------------------------------
-    samples = sample_int(n=1000, k=900, replace=True, p=p, use_numba=use_numba)  # very unlikely all are unique
+    samples = sample_int(n=1000, k=900, replace=True, p=p, accelerated=use_numba)  # very unlikely all are unique
 
     # --- assert ------------------------------------------
     assert len(set(samples)) < 900  # there are duplicates, which is expected with replacement
@@ -267,9 +267,9 @@ def test_sample_int_non_uniform_with_replacement_seed(n: int, k: int, use_numba:
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    samples_1 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=123)
-    samples_2 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=123)
-    samples_3 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=None)
+    samples_1 = sample_int(n, k, replace=True, p=p, accelerated=use_numba, seed=123)
+    samples_2 = sample_int(n, k, replace=True, p=p, accelerated=use_numba, seed=123)
+    samples_3 = sample_int(n, k, replace=True, p=p, accelerated=use_numba, seed=None)
 
     # --- assert ------------------------------------------
     np.testing.assert_array_equal(samples_1, samples_2)
@@ -292,7 +292,7 @@ def test_sample_int_non_uniform_with_replacement_probs(use_numba: bool, factor: 
     expected_mean = sum(i * p[i] for i in range(n))
 
     # --- act ---------------------------------------------
-    samples = sample_int(n=n, k=k, replace=True, p=p, use_numba=use_numba)
+    samples = sample_int(n=n, k=k, replace=True, p=p, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert 0.95 * expected_mean < np.mean(samples) < 1.05 * expected_mean
@@ -309,7 +309,7 @@ def test_sample_int_non_uniform_without_replacement_invariants_single(n: int, us
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    sample = sample_int(n, replace=False, p=p, use_numba=use_numba)
+    sample = sample_int(n, replace=False, p=p, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(sample, numbers.Integral)
@@ -334,7 +334,7 @@ def test_sample_int_non_uniform_without_replacement_invariants_multiple(n: int, 
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    samples = sample_int(n, k, replace=False, p=p, use_numba=use_numba)
+    samples = sample_int(n, k, replace=False, p=p, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(samples, np.ndarray)
@@ -362,9 +362,9 @@ def test_sample_int_non_uniform_without_replacement_seed(n: int, k: int, use_num
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    samples_1 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=123)
-    samples_2 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=123)
-    samples_3 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=None)
+    samples_1 = sample_int(n, k, replace=False, p=p, accelerated=use_numba, seed=123)
+    samples_2 = sample_int(n, k, replace=False, p=p, accelerated=use_numba, seed=123)
+    samples_3 = sample_int(n, k, replace=False, p=p, accelerated=use_numba, seed=None)
 
     # --- assert ------------------------------------------
     np.testing.assert_array_equal(samples_1, samples_2)
@@ -387,7 +387,7 @@ def test_sample_int_non_uniform_without_replacement_probs(use_numba: bool, facto
     expected_mean = sum(i * p[i] for i in range(n))  # approx. correct; this assumes replacement
 
     # --- act ---------------------------------------------
-    samples = sample_int(n=n, k=k, replace=False, p=p, use_numba=use_numba)
+    samples = sample_int(n=n, k=k, replace=False, p=p, accelerated=use_numba)
 
     # --- assert ------------------------------------------
     assert 0.95 * expected_mean < np.mean(samples) < 1.05 * expected_mean